GitStarCrawler/requirements.txt
# Core dependencies
requests==2.31.0
psycopg2-binary==2.9.9

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0
